requirements.txt
antiorm==1.2.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.5
attrs==21.4.0
autopep8==1.7.0
backcall==0.2.0
black==22.1.0
bleach==4.1.0
certifi==2022.9.24
cffi==1.15.0
charset-normalizer==2.1.1
click==8.0.3
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
db==0.1.1
db-sqlite3==0.0.1
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.5
dj-database-url==1.0.0
dj-rest-auth==2.2.5
Django==4.1.1
django-allauth==0.51.0
django-appconf==1.0.5
django-compressor==4.1
django-cors-headers==3.2.1
django-crispy-forms==1.14.0
django-debug-toolbar==3.7.0
django-filter==22.1
django-heroku==0.3.1
django-sass-processor==1.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.7.1
drf-yasg==1.21.4
entrypoints==0.4
executing==0.8.2
filelock==3.8.0
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
ipykernel==6.8.0
ipython==8.0.1
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
kaggle==1.5.12
libsass==0.21.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.5.10
nbconvert==6.4.1
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.8
numpy==1.22.3
oauthlib==3.2.1
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.4.1
prometheus-client==0.13.1
prompt-toolkit==3.0.26
psycopg2==2.9.3
pure-eval==0.2.2
pycodestyle==2.9.1
pycparser==2.21
Pygments==2.11.2
PyJWT==2.5.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-decouple==3.6
python-slugify==6.1.1
python3-openid==3.2.0
pytz==2022.4
pywin32==303
pywinpty==2.0.2
pyzmq==22.3.0
rcssmin==1.1.0
requests==2.28.1
requests-oauthlib==1.3.1
rjsmin==1.2.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
Send2Trash==1.8.0
six==1.16.0
sqlparse==0.4.3
stack-data==0.1.4
terminado==0.13.1
testpath==0.5.0
text-unidecode==1.3
toml==0.10.2
tomli==2.0.0
tornado==6.1
tqdm==4.63.0
traitlets==5.1.1
types-cryptography==3.3.23
tzdata==2022.4
uritemplate==4.1.1
urllib3==1.26.12
virtualenv==20.16.3
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==6.2.0
